pyproject.toml
[build-system]
requires = ["setuptools>=43.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xdufacool"
version = "0.9.0"
description = "A set of toolkit for faculty members in the Xidian University."
readme = "README.md"
requires-python = ">=3.9"
license = {file = "LICENSE"}
keywords = ["xidian", "university", "faculty", "toolkit"]
authors = [
  {name = "Fei Qi", email = "fred.qi@ieee.org"},
]
maintainers = [
  {name = "Fei Qi", email = "fred.qi@ieee.org"},
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python :: 3.8",
]
dependencies = [
  "arxiv-latex-cleaner",
  "dotenv",
  "jinja2",
  "mistletoe",
  "nbformat",
  "nbconvert",
  "numpy",
  "ipython",
  "jupytext",
  "openpyxl",
  "requests",
  "PyYAML",
  "pypandoc",
  "pysocks",
  "pygments",
  "python-docx",
  "tqdm",
  "docx-mailmerge2",
  "google-genai",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-cov",
]
gui = [
  "pyside6",
]

[project.urls]
homepage = "https://github.com/fredqi/xdufacool"

[project.scripts]
xdufacool = "xdufacool.cli:main"
dupcheck = "xdufacool.dupcheck:main"
subprep = "xdufacool.subprep:main"
beamer-translate = "xdufacool.beamer_translate:main"
# pdf2pptx = "xdufacool.pdf2pptx:pdf2pptx"
# syllabus = "xdufacool.syllabus:syllabus_helper"
# xduscore = "xdufacool.score_helper:xduscore"
# arxort = "xdufacool.organize_bib:organize_bib"

[tool.setuptools]
packages = ["xdufacool", "xdufacool.beamer"]
include-package-data = true

[tool.setuptools.package-data]
xdufacool = ["templates/*.tex.j2", "logging.yaml"]

[tool.pytest.ini_options]
testpaths = ["tests"] 